requirements.txt
matplotlib==3.3.4
numpy==1.19.2
opencv-python==4.5.1
pandas==1.2.3
pillow==8.1.2
tqdm==4.59.0